requirements.txt
# tensorflow==2.16.1
pandas
numpy
Pillow
matplotlib
scipy
scikit-learn
seaborn